% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postr.R
\docType{package}
\name{flex.poster-package}
\alias{flex.poster}
\alias{flex.poster-package}
\title{postr: Prepare reproducible posters using R Markdown}
\description{
Provides templates and helper functions to prepare R Markdown based
conference posters and produce output both in html for hosting and png for
printing. Relies on the flexdashboard package to easily define poster layout.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://rbcavanaugh.github.io/flex.poster}
  \item Report bugs at \url{https://github.com/rbcavanaugh/flex.poster/issues/}
}

}
\author{
\strong{Maintainer}: Rob Cavanaugh \email{rob.cavanaugh@pitt.edut}

}
